+++ docs/src/components/Header/Desktop.tsx
@@ -66,6 +66,14 @@ const Header: React.FC<Props> = (props) => (
             Slack
           </Button>
           <Button
+            href='https://forum.cube.dev/'
+            target='_blank'
+            className={styles.headerButton}
+          >
+            <Icon style={{ fontSize: '22px' }} type='discourse' />
+            Discourse
+          </Button>          
+          <Button
             href='https://github.com/cube-js/cube.js'
             target='_blank'
             className={styles.headerButton}