+++ examples/vue-query-builder/dashboard-app/src/main.js
@@ -3,12 +3,14 @@ import Chart from 'chart.js';
 import VueChartkick from 'vue-chartkick';
 import VueRouter from 'vue-router';
 import VueApollo from 'vue-apollo';
+import cubejs from '@cubejs-client/core';
 
 import App from './App.vue';
 import vuetify from './plugins/vuetify';
 import Explore from './pages/explore/Explore.vue';
 import Dashboard from './pages/dashboard/Dashboard.vue';
 import apolloClient from './graphql/client';
+import ChartContainer from '@/pages/test/ChartContainer';
 
 Vue.use(VueApollo);
 
@@ -16,7 +18,6 @@ const apolloProvider = new VueApollo({
   defaultClient: apolloClient
 });
 
-import cubejs from '@cubejs-client/core';
 const API_URL = 'https://ecom.cubecloudapp.dev';
 const CUBEJS_TOKEN =
   'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE1OTQ2NjY4OTR9.0fdi5cuDZ2t3OSrPOMoc3B1_pwhnWj4ZmM3FHEX7Aus';
@@ -29,6 +30,7 @@ const router = new VueRouter({
     { path: '/', component: Explore, props: { cubejsApi } },
     { path: '/explore', component: Explore, props: { cubejsApi } },
     { path: '/dashboard', component: Dashboard, props: { cubejsApi } },
+    { path: '/test', component: ChartContainer, props: { cubejsApi } },
   ]
 })