+++ packages/cubejs-backend-cloud/test/cloud.test.ts
@@ -1,0 +1,11 @@
+import { CubeCloudClient } from '../src/cloud';
+
+test('CubeCloudClient: constuctor', async () => {
+  const cubeCloudClient = new CubeCloudClient({
+    auth: '',
+    url: '',
+    deploymentId: '',
+    deploymentUrl: ''
+  });
+  expect(cubeCloudClient).not.toBeUndefined();
+});