+++ packages/cubejs-playground/src/PlaygroundQueryBuilder.tsx
@@ -54,6 +54,12 @@ export const frameworkChartLibraries = {
       title: 'ng2',
     },
   ],
+  vue: [
+    {
+      value: 'chartkick',
+      title: 'Chartkick',
+    },
+  ],
 };
 
 const playgroundActionUpdateMethods = (updateMethods, memberName) =>