+++ packages/cubejs-server-core/src/core/DevServer.ts
@@ -265,7 +265,8 @@ export class DevServer {
         await appContainer.ensureDependencies();
         this.cubejsServer.event('Dev Server Dashboard Npm Install Success');
 
-        fetcher.cleanup();
+        // todo: uncomment
+        // fetcher.cleanup();
       };
 
       if (this.applyTemplatePackagesPromise) {
@@ -332,7 +333,6 @@ export class DevServer {
 
     app.get('/playground/test-connection', catchErrors(async (req, res) => {
       const orchestratorApi = this.cubejsServer.getOrchestratorApi({
-        dataSource: req.query.dataSource || 'default',
         securityContext: null,
         authInfo: null,
         requestId: getRequestIdFromRequest(req),